Splits the library up into:
- cmd: handles commands, used by main
- audio: plain audio files
- video: plain video files
- movies: films
- tv: TV show episodes
- music: songs

Adds basic support for using [guessit](https://github.com/guessit-io/) to
extract information from (video) filenames. For books and audio, we'll need
something else (in which case I'll probably clone some of the guessit logic
and extend it for books and audio).

Also includes some starter tests for each of the libraries, because, go. And a
Makefile to run all the tests, since the `go test` command doesn't seem to
want to run sub-package tests (this is probably doable, I'm just too much of a
newbie).