Restructuring, adds template flag to scan...
...and adds all supported guessit fields. Also allows guessit to be run from
docker, in case you don't have it installed, and keeps a cache of the guesses
so it only has to be called once per filename (in the process), and some
simple debugging.

Music and Books are stubbed out, since guessit can't guess them - they'll be
added once the basic video features are there (TMDB/TVDB/IMDB lookups and
functions for shows, duplicates, etc).